pyproject.toml
[tool.poetry]
name = "okpostest"
version = "0.1.0"
description = ""
authors = ["KimDuHong <art970@naver.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "2.2.24"
drf-yasg = "^1.21.5"
django-debug-toolbar = "3.2.4"
drf-writable-nested = "^0.7.0"
djangorestframework = "3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
